+++ TestUtils.ts
@@ -55,7 +55,7 @@ const DEFAULT_GLOBAL_CONFIG: Config.GlobalConfig = {
   testNamePattern: '',
   testPathPattern: '',
   testResultsProcessor: null,
-  testSequencer: '',
+  testTimeout: 5000,
   updateSnapshot: 'none',
   useStderr: false,
   verbose: false,