+++ packages/jest-haste-map/src/lib/FSEventsWatcher.ts
@@ -11,14 +11,12 @@ import path from 'path';
 import {EventEmitter} from 'events';
 import anymatch from 'anymatch';
 import micromatch from 'micromatch';
-// eslint-disable-next-line no-unresolved
 import {Watcher} from 'fsevents';
 // @ts-ignore no types
 import walker from 'walker';
 
 let fsevents: (path: string) => Watcher;
 try {
-  // eslint-disable-next-line no-unresolved
   fsevents = require('fsevents');
 } catch (e) {
   // Optional dependency, only supported on Darwin.