+++ packages/jest-jasmine2/src/index.ts
@@ -34,6 +34,7 @@ async function jasmine2(
   const jasmine = jasmineFactory.create({
     process,
     testPath,
+    testTimeout: globalConfig.testTimeout,
   });
 
   const env = jasmine.getEnv();